# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ServRedi FSM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time-entries:
    get:
      operationId: listTimeEntries
      tags: [timeTracking]
      summary: List time entries
      parameters:
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: jobId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Time entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: createTimeEntry
      tags: [timeTracking]
      summary: Clock in — start a time entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeEntryBody"
      responses:
        "201":
          description: Created time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
...
  /time-entries/{id}/clock-out:
    patch:
      operationId: clockOut
      tags: [timeTracking]
      summary: Clock out — end a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutBody"
      responses:
        "200":
          description: Ended time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{id}/break:
    patch:
      operationId: updateBreak
      tags: [timeTracking]
      summary: Start or end a break
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBreakBody"
      responses:
        "200":
          description: Updated time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"

  # ── Quotes ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        technicianId:
          type: integer
        technicianName:
          type: string
        jobId:
          type: ["integer", "null"]
        jobTitle:
          type: ["string", "null"]
        taskId:
          type: ["integer", "null"]
        clockInAt:
          type: string
        clockOutAt:
          type: ["string", "null"]
        breakMinutes:
          type: integer
        totalMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - organizationId
        - technicianId
        - technicianName
        - clockInAt
        - breakMinutes
        - isActive
        - createdAt
        - updatedAt
...
    CreateTimeEntryBody:
      type: object
      properties:
        jobId:
          type: integer
        taskId:
          type: integer
        notes:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    ClockOutBody:
      type: object
      properties:
        notes:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    UpdateBreakBody:
      type: object
      properties:
        action:
          type: string
          enum: [start, end]
        breakMinutes:
          type: integer
      required:
        - action
...